# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9080","description":"Generated server url"}],"paths":{"/api/v1/bankTransaction":{"get":{"tags":["transaction-controller"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTransaction"}}}}}},"post":{"tags":["transaction-controller"],"operationId":"insert","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTransaction"}}}}}}},"/api/v1/bankAccount":{"get":{"tags":["account-controller"],"operationId":"findAllByName","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}},{"name":"title","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAccount"}}}}}},"post":{"tags":["account-controller"],"operationId":"insert_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAccount"}}}}}}},"/api/v1/bankAccount/{id}":{"get":{"tags":["account-controller"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"post":{"tags":["account-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAccount"}}}}}},"delete":{"tags":["account-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAccount"}}}}}}},"/api/v1/bankTransaction/{id}":{"get":{"tags":["transaction-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}}},"/api/v1/bankTransaction/transactions/{id}":{"get":{"tags":["transaction-controller"],"operationId":"getBalanceDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTransaction"}}}}}}},"/api/v1/bankBalance":{"get":{"tags":["balance-controller"],"operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Balance"}}}}}}}},"/api/v1/bankBalance/{id}":{"get":{"tags":["balance-controller"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Balance"}}}}}}},"/api/v1/bankBalance/balance/{id}":{"get":{"tags":["balance-controller"],"operationId":"getBalanceDetails_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBalance"}}}}}}},"/api/v1/bankBalance/auth":{"get":{"tags":["balance-controller"],"operationId":"getAuthenticatedUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAccount"}}}}}}},"/api/v1/bankAccount/email/{email}":{"get":{"tags":["account-controller"],"operationId":"findByEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOptionalAccount"}}}}}}},"/api/v1/bankAccount/auth":{"get":{"tags":["account-controller"],"operationId":"getAuthenticatedUser_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAccount"}}}}}}},"/api/v1/bankAccount/all":{"get":{"tags":["account-controller"],"operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}}},"/login":{"post":{"tags":["login-endpoint"],"requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"OK"},"403":{"description":"Forbidden"}}}}},"components":{"schemas":{"Transaction":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"integer","format":"int64"},"description":{"type":"string"},"transactionType":{"type":"string"},"date":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"}}},"ApiResponseTransaction":{"type":"object","properties":{"content":{"$ref":"#/components/schemas/Transaction"}}},"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"type":"string"},"roles":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"}}},"ApiResponseAccount":{"type":"object","properties":{"content":{"$ref":"#/components/schemas/Account"}}},"ApiResponseListTransaction":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}},"Balance":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"integer","format":"int64"},"lastTransaction":{"type":"string"},"date":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"}}},"ApiResponseBalance":{"type":"object","properties":{"content":{"$ref":"#/components/schemas/Balance"}}},"ApiResponseListAccount":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}},"ApiResponseOptionalAccount":{"type":"object","properties":{"content":{"$ref":"#/components/schemas/Account"}}}}}}